+++ package.json
@@ -24,7 +24,8 @@
     "vue-router": "^3.1.3",
     "vuetify": "^2.0.0",
     "vuetify-loader": "^1.2.2",
-    "vuex": "^3.0.1"
+    "vuex": "^3.0.1",
+    "vuex-persist": "^2.2.0"
   },
   "devDependencies": {
     "@mdi/font": "^4.1.95",
