+++ src/components/Generator/Ingress.vue
@@ -57,12 +57,12 @@ export default {
   data () {
     return {
       form: {
-        ingressname: this.$store.getters['get_ingressname'],
-        serviceport: this.$store.getters['get_serviceport'],
-        servicename: this.$store.getters['get_servicename'],
-        domainname: this.$store.getters['get_domainname'],
-        subdomain: this.$store.getters['get_subdomain'],
-        namespace: this.$store.getters['get_namespace']
+        ingressname: this.$store.getters['generator/get_ingressname'],
+        serviceport: this.$store.getters['generator/get_serviceport'],
+        servicename: this.$store.getters['generator/get_servicename'],
+        domainname: this.$store.getters['generator/get_domainname'],
+        subdomain: this.$store.getters['generator/get_subdomain'],
+        namespace: this.$store.getters['generator/get_namespace']
       },
       domains: [{ text: 'choose...', value: '' }, 'demo.datexis.com', 'app.datexis.com', 'internal.datexis.com', 'api.datexis.com'],
       options: ['enable CORS', 'restriction to beuth network', 'cors-allow-origin'],
@@ -135,7 +135,7 @@ export default {
     },
     async checkHostName () {
       let hostname = this.form.subdomain + '.' + this.form.domainname
-      const validation = await this.$store.dispatch('validate_hostname', hostname)
+      const validation = await this.$store.dispatch('generator/validate_hostname', hostname)
       return validation
     }
   },
