+++ src/main.js
@@ -1,18 +1,29 @@
 import Vue from 'vue'
 import App from './App.vue'
-import router from './router'
+import VuexPersist from 'vuex-persist';
 import store from './store'
-import '@/assets/css/tailwind.css'
 import logger from '@/plugins/logger'
 import vuetify from './plugins/vuetify'
+import router from './router'
 import { BootstrapVue, IconsPlugin } from 'bootstrap-vue'
 import GAuth from 'vue-google-oauth2'
+import '@/assets/css/tailwind.css'
 
 const gauthOption = {
   clientId: '516982342717-6ri9qn82ucchr13ftdec279bm2if82n0.apps.googleusercontent.com',
   scope: 'profile email',
   prompt: 'select_account'
 }
+
+const vuexLocalStorage = new VuexPersist({
+  key: 'vuex', // The key to store the state on in the storage provider.
+  storage: window.localStorage, // or window.sessionStorage or localForage
+  // Function that passes the state and returns the state with only the objects you want to store.
+  reducer: state => ({
+    keepThisModule: state.keepThisModule,
+    keepThisModuleToo: state.keepThisModuleToo
+  })
+})
 
 Vue.use(GAuth, gauthOption)
 Vue.use(BootstrapVue)
@@ -27,3 +38,5 @@ new Vue({
   vuetify,
   render: h => h(App)
 }).$mount('#app')
+
+export default vuexLocalStorage
